revert modifying BigDecimal into Int

+++ ocpi-toolkit-2.2.1/src/main/kotlin/com/izivia/ocpi/toolkit/modules/cdr/domain/CdrDimension.kt
@@ -10,5 +10,5 @@ import java.math.BigDecimal
 @Partial
 data class CdrDimension(
     val type: CdrDimensionType,
-    val volume: Int
+    val volume: BigDecimal
 )
